refactor: 壊変経路を取り扱う処理をDecaySetクラスに切り出す
