+++ FlexID.Calc/InputDataReader_OIR.cs
@@ -804,149 +804,12 @@ public class InputDataReader_OIR : InputDataReaderBase
     }
 
     /// <summary>
-    /// 対象核種から始まる崩壊系列を構成する子孫核種を列挙する。
-    /// </summary>
-    /// <param name="root"></param>
-    /// <returns>rootを含まない、子孫核種のみの配列。</returns>
-    private NuclideData[] GetDecayNuclides(NuclideData root)
-    {
-        var results = new List<NuclideData> { root };
-
-    Lagain:
-        var count = results.Count;
-        for (int i = 0; i < count; i++)
-        {
-            // 系列を構成する核種の娘を追加する。
-            var nuclide = results[i];
-            var daughters = nuclide.Branches.Select(b => b.Daughter);
-            results.AddRange(daughters.Where(d => !results.Contains(d)));
-        }
-        // 系列を構成する核種が増えなくなるまで繰り返す。
-        if (results.Count > count)
-            goto Lagain;
-
-        // rootを除いた、子孫核種のみの配列を返す。
-        return results.Skip(1).ToArray();
-    }
-
-    /// <summary>
-    /// 対象コンパートメントの核種から始まる崩壊系列の情報を保持する。
-    /// </summary>
-    private class DecayChain
-    {
-        /// <summary>
-        /// コンストラクタ。
-        /// </summary>
-        /// <param name="organRoot"></param>
-        /// <param name="decayNuclides">子孫核種の配列。</param>
-        public DecayChain(Organ organRoot, NuclideData[] decayNuclides)
-        {
-            RootCompartment = organRoot;
-            DecayNuclides = decayNuclides;
-            DecayCompartments = new Organ[decayNuclides.Length];
-        }
-
-        public Organ RootCompartment;
-
-        public NuclideData[] DecayNuclides;
-
-        public Organ[] DecayCompartments;
-
-        /// <summary>
-        /// 崩壊系列を構成する全ての移行経路を取得する。
-        /// </summary>
-        /// <returns></returns>
-        public IReadOnlyList<(NuclideData Parent, NuclideData Daughter)> GetDecayTransfers()
-        {
-            var nuclides = DecayNuclides.Prepend(RootCompartment.Nuclide);
-
-            var results = new List<(NuclideData Parent, NuclideData Daughter)>();
-            foreach (var parent in nuclides)
-            {
-                foreach (var daughter in DecayNuclides)
-                {
-                    if (parent == daughter)
-                        continue;
-                    if (parent.Branches.Any(b => b.Daughter == daughter))
-                        results.Add((parent, daughter));
-                }
-            }
-            return results;
-        }
-
-        /// <summary>
-        /// 指定の子孫核種に対応する壊変コンパートメントを追加する。
-        /// </summary>
-        /// <param name="data"></param>
-        /// <param name="progeny"></param>
-        /// <returns>追加した壊変コンパートメント</returns>
-        public Organ AddDecayCompartment(InputData data, NuclideData progeny)
-        {
-            var organFrom = RootCompartment;
-            var nuclideFrom = organFrom.Nuclide;
-
-            var index = data.Organs.Count;
-            var organDecay = new Organ
-            {
-                Nuclide = progeny,
-                ID = index + 1,
-                Index = index,
-                Name = $"Decay-{nuclideFrom.Name}/{organFrom.Name}",
-                Func = OrganFunc.acc,
-                BioDecay = 1.0,     // accは後で設定する。
-                IsDecayCompartment = true,
-            };
-            data.Organs.Add(organDecay);
-
-            var sourceRegion = organFrom.SourceRegion;
-            if (sourceRegion != null)
-            {
-                // 親核種のコンパートメントに設定された線源領域が、
-                // 子孫核種の動態モデルおいても明示的に設定されているかどうかを調べる。
-                var explicitUsed = data.Organs.Where(o => o.Nuclide == progeny)
-                                              .Any(o => o.SourceRegion == sourceRegion);
-
-                // 明示的に設定されていない場合は、ambiguous compartmentを'Other'に割り当てる。
-                if (!explicitUsed)
-                    sourceRegion = "Other";
-
-                organDecay.SourceRegion = sourceRegion;
-            }
-
-            return organDecay;
-        }
-    }
-
-    /// <summary>
     /// 全ての移行経路を定義する。
     /// </summary>
     /// <param name="data"></param>
     private void DefineTransfers(InputData data)
     {
-        var decayPaths = new List<(Organ from, Organ to, bool hasCoeff)>();
-        var decayChains = new Dictionary<Organ, DecayChain>();
-        var decayNuclides = data.Nuclides.ToDictionary(n => n, n => GetDecayNuclides(n));
-
-        // organFromから始まる崩壊系列の情報を取得する。
-        DecayChain GetDecayChain(Organ organFrom)
-        {
-            var funcFrom = organFrom.Func;
-            if (funcFrom != OrganFunc.acc && funcFrom != OrganFunc.exc)
-                return null;
-            if (organFrom.IsDecayCompartment)
-                return null;
-
-            var progenies = decayNuclides[organFrom.Nuclide];
-            if (progenies.Length == 0)
-                return null;
-
-            if (!decayChains.TryGetValue(organFrom, out var decayChain))
-            {
-                decayChain = new DecayChain(organFrom, progenies);
-                decayChains[organFrom] = decayChain;
-            }
-            return decayChain;
-        }
+        var decaySet = new DecaySet(data.Nuclides, errors, data.Organs);
 
         foreach (var nuclide in nuclides)
         {
@@ -1080,18 +943,6 @@ public class InputDataReader_OIR : InputDataReaderBase
 
                 if (isDecayPath)
                 {
-                    // 分岐比が不明な壊変経路は定義できない。
-                    if (!decayNuclides[nuclideFrom].Contains(nuclideTo))
-                        errors.AddError(lineNum, $"There is no decay path from {nuclideFrom.Name} to {nuclideTo.Name}.");
-
-                    var paths = decayPaths.Where(path => path.from == organFrom);
-
-                    // organFromから、同じ子孫核種への2つ以上の壊変経路は定義できない。
-                    if (paths.Any(path => path.to.Nuclide == nuclideTo))
-                        errors.AddError(lineNum, $"Multiple decay paths from {funcFrom} '{from}' to nuclide '{nuclideTo.Name}'.");
-
-                    decayPaths.Add((organFrom, organTo, hasCoeff));
-
                     // 次のような壊変経路を、
                     //   Parent/organFrom --(coeff)--> Progeny_i/organTo  ①移行速度あり
                     //   Parent/organFrom -----------> Progeny_i/organTo  ②移行速度なし
@@ -1111,41 +962,12 @@ public class InputDataReader_OIR : InputDataReaderBase
                     //   ↓ ↓
                     //   Progeny_N/organDecay
 
-                    var decayChain = GetDecayChain(organFrom);
-
-                    var decayIndex = Array.IndexOf(decayChain.DecayNuclides, nuclideTo);
-                    if (decayIndex == -1)
-                    {
-                        var rootNuc = decayChain.RootCompartment.Nuclide.Name;
-                        errors.AddError(lineNum, $"Cannot find progeny nuclide '{nuclideTo.Name}' in decay chain starts from '{rootNuc}'.");
+                    var organDecay = decaySet.AddDecayPath(lineNum, organFrom, organTo, coeff);
+                    if (organDecay is null)
                         continue;
-                    }
 
-                    ref var organDecay = ref decayChain.DecayCompartments[decayIndex];
-                    if (organDecay != null)
-                    {
-                        if (hasCoeff)
-                            errors.AddError(lineNum, $"Decay compartment is already set.");
-                        else
-                            errors.AddError(lineNum, $"Decay compartment '{to}' conflicts with the implicitly deined one.");
-                        continue;
-                    }
-
-                    if (hasCoeff)
-                    {
-                        organDecay = decayChain.AddDecayCompartment(data, nuclideTo);
-
-                        // 以降の処理を核種が同じコンパートメント間organDecay -> organToの経路設定にすり替える。
-                        organFrom = organDecay;
-                    }
-                    else
-                    {
-                        // organToを崩壊系列でorganDecayが占める位置に設定する。
-                        organDecay = organTo;
-
-                        // transfersCorrectには核種が異なるコンパートメント間の壊変経路を追加しない。
-                        continue;
-                    }
+                    // 以降の処理を核種が同じコンパートメント間organDecay -> organToの経路設定にすり替える。
+                    organFrom = organDecay;
                 }
 
                 if (coeff is decimal coeff_v)
@@ -1240,56 +1062,10 @@ public class InputDataReader_OIR : InputDataReaderBase
         errors.RaiseIfAny();
 
         // 核種が異なるコンパートメントへの流入経路と、移行割合を設定する。
-        foreach (var decayChain in decayChains.Values)
-        {
-            DefineDecayTransfers(data, decayChain);
-        }
+        decaySet.DefineDecayTransfers();
 
         // 移行経路の定義にエラーがないことを確定する。
         errors.RaiseIfAny();
-    }
-
-    /// <summary>
-    /// 核種が異なるコンパートメント間の移行経路を定義する。
-    /// </summary>
-    /// <param name="data"></param>
-    private void DefineDecayTransfers(InputData data, DecayChain chain)
-    {
-        var organFrom = chain.RootCompartment;
-        var nuclideFrom = organFrom.Nuclide;
-        var decayTransfers = chain.GetDecayTransfers();
-        var decayCompartments = chain.DecayCompartments;
-
-        // 崩壊系列を構成するコンパートメント間の壊変経路を追加定義する。
-        // 壊変経路は一本道ではなく、nuclideFromから始まる有効非巡回グラフ(DAG)を構成する点に注意。
-        foreach (var path in decayTransfers)
-        {
-            var from = path.Parent == nuclideFrom ? organFrom
-                   : decayCompartments.FirstOrDefault(o => o?.Nuclide == path.Parent);
-            var to = decayCompartments.FirstOrDefault(o => o?.Nuclide == path.Daughter);
-
-            // 壊変経路が既に設定されている＝インプットで明示的に定義されている場合は何もしない。
-            if (from != null && to != null && to.Inflows.Any(inflow => inflow.Organ == from))
-                continue;
-
-            if (from is null)
-                from = chain.AddDecayCompartment(data, path.Parent);
-            if (to is null)
-                to = chain.AddDecayCompartment(data, path.Daughter);
-
-            var branch = from.Nuclide.Branches.First(b => b.Daughter == to.Nuclide);
-
-            to.Inflows.Add(new Inflow
-            {
-                ID = from.ID,
-
-                // 壊変経路では、親からの分岐比を移行割合としてとする。
-                Rate = branch.Fraction,
-
-                // 流入経路から流入元臓器の情報を直接引くための参照を設定する。
-                Organ = from,
-            });
-        }
     }
 
     /// <summary>
